+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Support for PHP 5.6
 
 ### Added
-- New installer type - `oxideshop-component`, which will trigger service registration to OXID eShop DI container.
+- Backwards compatibility break: New installer type - `oxideshop-component`, which will trigger service registration to newly introduced OXID eShop DI container.
 
 ### Changed
 - Exclude non-essential files from dist package [PR-12](https://github.com/OXID-eSales/oxideshop_composer_plugin/pull/12)