+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 ### Deprecated
+- Class OxidEsales\ComposerPlugin\Installer\Package\ModulePackageInstaller
 
 ### Removed